+++ backend/services/sql-test/src/main/java/io/metersphere/sql/service/common/UserConnectionInfoService.java
@@ -3,8 +3,8 @@ package io.metersphere.sql.service.common;
 import io.metersphere.sql.domain.UserConnectionInfo;
 
 public interface UserConnectionInfoService {
-    Integer createUserConnectionInfo(UserConnectionInfo userConnectionInfo);
+    Long createUserConnectionInfo(UserConnectionInfo userConnectionInfo);
 
-    UserConnectionInfo getUserConnectionInfo(Integer id);
+    UserConnectionInfo getUserConnectionInfo(Long id);
 
 }
